Add Swarm.Language.LSP and rename --check into --format
