+++ swarm.cabal
@@ -60,6 +60,7 @@ library
                       Swarm.Language.Pretty
                       Swarm.Language.Typecheck
                       Swarm.Language.Elaborate
+                      Swarm.Language.LSP
                       Swarm.Language.Pipeline
                       Swarm.Language.Pipeline.QQ
                       Swarm.Game.CEK
@@ -99,6 +100,7 @@ library
                       hsnoise                       >= 0.0.2 && < 0.1,
                       lens                          >= 4.19 && < 5.1,
                       linear                        >= 1.21.6 && < 1.22,
+                      lsp                           >= 1.2 && < 1.3,
                       mtl                           >= 2.2.2 && < 2.3,
                       minimorph                     >= 0.3 && < 0.4,
                       murmur3                       >= 1.0.4 && < 1.1,
@@ -125,10 +127,9 @@ library
 executable swarm
     import:           stan-config, common
     main-is:          Main.hs
-    build-depends:    lsp                            >= 1.2 && < 1.3,
+    build-depends:    optparse-applicative          >= 0.16 && < 0.17,
                       -- Import shared with the library don't need bounds
                       base,
-                      lens,
                       text,
                       swarm
     hs-source-dirs:   app
